Release 7.4.0 (#15)
* Update project to latest tag saceriam-jboss-7.4.0

* Update SNAPSHOT version

* [maven-release-plugin] prepare release saceriam-jboss-7.4.0

* [maven-release-plugin] prepare for next development iteration

---------

+++ SacerIam-ejb/src/main/java/it/eng/saceriam/job/replicaUtenti/ejb/ReplicaUtentiEjb.java
@@ -611,7 +611,7 @@ public class ReplicaUtentiEjb {
 		abilDatiMap.put(dato.getIdOrganizIam(), abil);
 	    }
 	    TipoDatoAbil tipoDatoAbil = new TipoDatoAbil();
-	    tipoDatoAbil.setIdTipoDatoApplic(dato.getIdTipoDatoApplic().intValue());
+	    tipoDatoAbil.setIdTipoDatoApplic(dato.getIdTipoDatoApplic().longValue());
 	    tipoDatoAbil.setNmClasseTipoDato(dato.getNmClasseTipoDato());
 
 	    abil.add(tipoDatoAbil);
